refactor: file structure organisation
Relates #256
