+++ components/icons/ManIcon.tsx
@@ -1,4 +1,4 @@
-import { IconPropType } from '../filter/QuickBrowse';
+import { IconPropType } from '../search/filter/QuickBrowse';
 
 const ManIcon: React.FC<IconPropType> = ({ subcategory }) => {
   return (
